Move responses mocking helpers to a shared lib
