+++ codex-rs/core/tests/common/Cargo.toml
@@ -11,3 +11,4 @@ codex-core = { path = "../.." }
 serde_json = "1"
 tempfile = "3"
 tokio = { version = "1", features = ["time"] }
+wiremock = "0.6"
